refactor: extract experience and updater modules
